backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
python-multipart==0.0.9
pydantic==2.10.6
pydantic-settings==2.7.1
google-cloud-speech==2.30.0
google-cloud-aiplatform==1.79.0
